Daily 15.03.2021
In the process of implementing buds and lateral branching.  Bud update is handled the same as node update is (overwrite every frame), which does not currently register if a bud has turned into a branch. Rather buds and nodes should be coupled by a unique attribute, which allows for a node to be stopped from overwriting a bud after it has opened