Light field
a voxel light field is created and initialized with a directional light vector. Basic "shadow" created by intersecting each voxels position along the reversed light vector with the skinned model if the tree. This produces a padding around the tree. Currently constrained by distance to partly resolve the problem. Should be tested with dot product, and if the geometry of the tree was watertight.